Migrate Plan into it's own Planner class.
