+++ reasoner/resolution/ResolverManager.java
@@ -29,7 +29,6 @@ import grakn.core.logic.resolvable.Resolvable;
 import grakn.core.logic.resolvable.Retrievable;
 import grakn.core.pattern.Conjunction;
 import grakn.core.pattern.equivalence.AlphaEquivalence;
-import grakn.core.pattern.variable.Variable;
 import grakn.core.reasoner.resolution.framework.ResolutionAnswer;
 import grakn.core.reasoner.resolution.resolver.ConcludableResolver;
 import grakn.core.reasoner.resolution.resolver.ResolvableResolver;
@@ -41,19 +40,14 @@ import graql.lang.pattern.variable.Reference;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.Comparator;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 import java.util.Set;
 import java.util.function.Consumer;
 import java.util.function.Function;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 import static grakn.common.collection.Collections.set;
 import static grakn.core.common.exception.ErrorMessage.Internal.ILLEGAL_STATE;
@@ -85,7 +79,7 @@ public class ResolverManager {
     }
 
     public List<MappedActor> planAndRegister(Set<Resolvable> resolvables) {
-        return iterate(new Plan(resolvables, conceptMgr, logicMgr).plan()).map(this::registerResolvable).toList();
+        return iterate(new Planner(conceptMgr, logicMgr).plan(resolvables)).map(this::registerResolvable).toList();
     }
 
     public Actor<RuleResolver> registerRule(Rule rule) {
@@ -157,136 +151,4 @@ public class ResolverManager {
         }
     }
 
-    static class Plan {
-        private final List<Resolvable> plan;
-        private final Map<Resolvable, Set<Variable>> dependencies;
-        private final ConceptManager conceptMgr;
-        private final LogicManager logicMgr;
-        private final Set<Variable> varsAnswered;
-        private final Set<Resolvable> remaining;
-
-        Plan(Set<Resolvable> resolvables, ConceptManager conceptMgr, LogicManager logicMgr) {
-            this.conceptMgr = conceptMgr;
-            this.logicMgr = logicMgr;
-            assert resolvables.size() > 0;
-            this.plan = new ArrayList<>();
-            this.varsAnswered = new HashSet<>();
-            this.dependencies = dependencies(resolvables);
-            this.remaining = new HashSet<>(resolvables);
-            computePlan();
-            assert plan.size() == resolvables.size();
-            assert set(plan).equals(resolvables);
-        }
-
-        private void add(Resolvable resolvable) {
-            plan.add(resolvable);
-            varsAnswered.addAll(namedVariables(resolvable));
-            remaining.remove(resolvable);
-        }
-
-        List<Resolvable> plan() {
-            return plan;
-        }
-
-        private void computePlan() {
-            while (remaining.size() != 0) {
-                Optional<Concludable> concludable;
-                Optional<Resolvable> retrievable;
-
-                // Retrievable where:
-                // all of it's dependencies are already satisfied,
-                // which will answer the most variables
-                retrievable = mostUnansweredVars(dependenciesSatisfied(connected(remaining.stream().filter(Resolvable::isRetrievable))));
-                if (retrievable.isPresent()) {
-                    add(retrievable.get());
-                    continue;
-                }
-
-                // Concludable where:
-                // all of it's dependencies are already satisfied,
-                // which has the least applicable rules,
-                // and of those the least unsatisfied variables
-                concludable = fewestRules(dependenciesSatisfied(connected(remaining.stream().filter(Resolvable::isConcludable))));
-                if (concludable.isPresent()) {
-                    add(concludable.get());
-                    continue;
-                }
-
-                // Retrievable where:
-                // all of it's dependencies are already satisfied (should be moot),
-                // it can be disconnected
-                // which will answer the most variables
-                retrievable = mostUnansweredVars(dependenciesSatisfied(remaining.stream().filter(Resolvable::isRetrievable)));
-                if (retrievable.isPresent()) {
-                    add(retrievable.get());
-                    continue;
-                }
-
-                // Concludable where:
-                // it can be disconnected
-                // all of it's dependencies are already satisfied,
-                // which has the least applicable rules,
-                // and of those the least unsatisfied variables
-                concludable = fewestRules(dependenciesSatisfied(remaining.stream().filter(Resolvable::isConcludable)));
-                if (concludable.isPresent()) {
-                    add(concludable.get());
-                    continue;
-                }
-
-                // Concludable where:
-                // it can be disconnected
-                // all of it's dependencies are NOT already satisfied,
-                // which has the least applicable rules,
-                // and of those the least unsatisfied variables
-                concludable = fewestRules(remaining.stream().filter(Resolvable::isConcludable));
-                if (concludable.isPresent()) {
-                    add(concludable.get());
-                    continue;
-                }
-                throw GraknException.of(ILLEGAL_STATE);
-            }
-        }
-
-        private Stream<Resolvable> dependenciesSatisfied(Stream<Resolvable> resolvableStream) {
-            return resolvableStream.filter(c -> varsAnswered.containsAll(dependencies.get(c)));
-        }
-
-        private Stream<Resolvable> connected(Stream<Resolvable> resolvableStream) {
-            return resolvableStream.filter(r -> !Collections.disjoint(namedVariables(r), varsAnswered));
-        }
-
-        private Optional<Concludable> fewestRules(Stream<Resolvable> resolvableStream) {
-            // TODO Tie-break for Concludables with the same number of applicable rules
-            return resolvableStream.map(Resolvable::asConcludable)
-                    .min(Comparator.comparingInt(c -> c.getApplicableRules(conceptMgr, logicMgr).toSet().size()));
-        }
-
-        private Optional<Resolvable> mostUnansweredVars(Stream<Resolvable> resolvableStream) {
-            return resolvableStream.max(Comparator.comparingInt(r -> iterate(namedVariables(r))
-                    .filter(var -> !varsAnswered.contains(var)).toSet().size()));
-        }
-
-        /**
-         * Determine the resolvables that are dependent upon the generation of each variable
-         */
-        private Map<Resolvable, Set<Variable>> dependencies(Set<Resolvable> resolvables) {
-            Map<Resolvable, Set<Variable>> deps = new HashMap<>();
-            Set<Variable> generatedVars = iterate(resolvables).filter(Resolvable::isConcludable)
-                    .map(Resolvable::asConcludable).map(Concludable::generating).toSet();
-            for (Resolvable resolvable : resolvables) {
-                for (Variable v : namedVariables(resolvable)) {
-                    deps.putIfAbsent(resolvable, new HashSet<>());
-                    if (generatedVars.contains(v) && !(resolvable.isConcludable() && resolvable.asConcludable().generating().equals(v))) {
-                        // TODO Should this rule the Resolvable out if generates it's own dependency?
-                        deps.get(resolvable).add(v);
-                    }
-                }
-            }
-            return deps;
-        }
-
-        private Set<Variable> namedVariables(Resolvable resolvable) {
-            return iterate(resolvable.conjunction().variables()).filter(var -> var.reference().isName()).toSet();
-        }
-    }
 }
