+++ test/integration/reasoner/resolution/PlannerTest.java
@@ -47,7 +47,7 @@ import static grakn.common.collection.Collections.list;
 import static grakn.common.collection.Collections.set;
 import static junit.framework.TestCase.assertEquals;
 
-public class ResolverManagerTest {
+public class PlannerTest {
 
     private static Path directory = Paths.get(System.getProperty("user.dir")).resolve("resolver-manager-test");
     private static String database = "resolver-manager-test";
@@ -86,7 +86,7 @@ public class ResolverManagerTest {
         Retrievable retrievable = new Retrievable(parse("{ $c($b); }"));
 
         Set<Resolvable> resolvables = set(concludable, retrievable);
-        List<Resolvable> plan = new ResolverManager.Plan(resolvables, conceptMgr, logicMgr).plan();
+        List<Resolvable> plan = new Planner(conceptMgr, logicMgr).plan(resolvables);
         assertEquals(list(concludable, retrievable), plan);
     }
 
@@ -100,7 +100,7 @@ public class ResolverManagerTest {
 
         Set<Resolvable> resolvables = set(concludable, retrievable);
 
-        List<Resolvable> plan = new ResolverManager.Plan(resolvables, conceptMgr, logicMgr).plan();
+        List<Resolvable> plan = new Planner(conceptMgr, logicMgr).plan(resolvables);
         assertEquals(list(retrievable, concludable), plan);
     }
 
@@ -120,7 +120,7 @@ public class ResolverManagerTest {
 
         Set<Resolvable> resolvables = set(retrievable, retrievable2, concludable);
 
-        List<Resolvable> plan = new ResolverManager.Plan(resolvables, conceptMgr, logicMgr).plan();
+        List<Resolvable> plan = new Planner(conceptMgr, logicMgr).plan(resolvables);
         assertEquals(list(retrievable, concludable, retrievable2), plan);
     }
 
@@ -140,7 +140,7 @@ public class ResolverManagerTest {
 
         Set<Resolvable> resolvables = set(retrievable, retrievable2, concludable);
 
-        List<Resolvable> plan = new ResolverManager.Plan(resolvables, conceptMgr, logicMgr).plan();
+        List<Resolvable> plan = new Planner(conceptMgr, logicMgr).plan(resolvables);
         assertEquals(list(retrievable2, concludable, retrievable), plan);
     }
 
@@ -151,7 +151,7 @@ public class ResolverManagerTest {
 
         Set<Resolvable> resolvables = set(concludable, concludable2);
 
-        List<Resolvable> plan = new ResolverManager.Plan(resolvables, conceptMgr, logicMgr).plan();
+        List<Resolvable> plan = new Planner(conceptMgr, logicMgr).plan(resolvables);
         assertEquals(list(concludable, concludable2), plan);
     }
 
@@ -170,7 +170,7 @@ public class ResolverManagerTest {
         Concludable concludable2 = Concludable.create(parse("{ $e($c, $p2) isa employment; }")).iterator().next();
 
         Set<Resolvable> resolvables = set(retrievable, retrievable2, concludable, concludable2);
-        List<Resolvable> plan = new ResolverManager.Plan(resolvables, conceptMgr, logicMgr).plan();
+        List<Resolvable> plan = new Planner(conceptMgr, logicMgr).plan(resolvables);
 
         assertEquals(list(retrievable, concludable, retrievable2, concludable2), plan);
     }
@@ -181,7 +181,7 @@ public class ResolverManagerTest {
         Concludable concludable2 = Concludable.create(parse("{ $b has $a; }")).iterator().next();
 
         Set<Resolvable> resolvables = set(concludable, concludable2);
-        List<Resolvable> plan = new ResolverManager.Plan(resolvables, conceptMgr, logicMgr).plan();
+        List<Resolvable> plan = new Planner(conceptMgr, logicMgr).plan(resolvables);
 
         assertEquals(2, plan.size());
         assertEquals(set(concludable, concludable2), set(plan));
@@ -193,7 +193,7 @@ public class ResolverManagerTest {
         Concludable concludable2 = Concludable.create(parse("{ $b($a); }")).iterator().next();
 
         Set<Resolvable> resolvables = set(concludable, concludable2);
-        List<Resolvable> plan = new ResolverManager.Plan(resolvables, conceptMgr, logicMgr).plan();
+        List<Resolvable> plan = new Planner(conceptMgr, logicMgr).plan(resolvables);
 
         assertEquals(2, plan.size());
         assertEquals(set(concludable, concludable2), set(plan));
@@ -205,7 +205,7 @@ public class ResolverManagerTest {
         Concludable concludable2 = Concludable.create(parse("{ $c($d); }")).iterator().next();
 
         Set<Resolvable> resolvables = set(concludable, concludable2);
-        List<Resolvable> plan = new ResolverManager.Plan(resolvables, conceptMgr, logicMgr).plan();
+        List<Resolvable> plan = new Planner(conceptMgr, logicMgr).plan(resolvables);
 
         assertEquals(2, plan.size());
         assertEquals(set(concludable, concludable2), set(plan));
@@ -229,7 +229,7 @@ public class ResolverManagerTest {
         Concludable concludable2 = Concludable.create(parse("{ $c($b) isa friendship; }")).iterator().next();
 
         Set<Resolvable> resolvables = set(concludable, concludable2);
-        List<Resolvable> plan = new ResolverManager.Plan(resolvables, conceptMgr, logicMgr).plan();
+        List<Resolvable> plan = new Planner(conceptMgr, logicMgr).plan(resolvables);
 
         assertEquals(0, concludable.getApplicableRules(conceptMgr, logicMgr).toList().size());
         assertEquals(1, concludable2.getApplicableRules(conceptMgr, logicMgr).toList().size());
